apps/analysis/pyproject.toml
[project]
name = "chesslens-analysis"
version = "0.1.0"
description = "ChessLens analysis service — Stockfish engine + AI coaching"
requires-python = ">=3.11"
license = {text = "AGPL-3.0"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-chess>=1.999",
    "openai>=1.60.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
